+++ src/ts/components/dates/DateTimePicker.tsx
@@ -12,6 +12,7 @@ import { StylesApiProps } from "props/styles";
 import React, { useState } from "react";
 import { isDisabled } from "../../utils/dates";
 import { setPersistence, getLoadingState } from "../../utils/dash3";
+import { resolveProp } from "../../utils/prop-functions"
 
 interface Props
     extends DashBaseProps,
@@ -34,8 +35,8 @@ interface Props
     submitButtonProps?: Omit<ActionIconProps, "n_click"> & any;
     /** Determines whether seconds input should be rendered */
     withSeconds?: boolean;
-    /** Specifies days that should be disabled */
-    disabledDates?: string[];
+    /** Specifies days that should be disabled.  Either a list of dates or a function. See https://www.dash-mantine-components.com/functions-as-props */
+    disabledDates?: any;
     /** An integer that represents the number of times that this element has been submitted */
     n_submit?: number;
     /** Debounce time in ms */
@@ -89,7 +90,7 @@ const DateTimePicker = ({
             onKeyDown={handleKeyDown}
             minDate={minDate}
             maxDate={maxDate}
-            excludeDate={isExcluded}
+            excludeDate={Array.isArray(disabledDates)? isExcluded : resolveProp(disabledDates)}
             {...others}
         />
     );