+++ src/ts/components/dates/MonthPickerInput.tsx
@@ -16,8 +16,6 @@ interface Props
     StylesApiProps {
     /** Dayjs format to display input value, "MMMM D, YYYY" by default  */
     valueFormat?: string;
-    /** Specifies days that should be disabled */
-    disabledDates?: string[];
     /** An integer that represents the number of times that this element has been submitted */
     n_submit?: number;
     /** Debounce time in ms */
@@ -34,7 +32,6 @@ const MonthPickerInput = ({
     debounce = false,
     minDate,
     maxDate,
-    disabledDates,
     popoverProps,
     persistence,
     persisted_props,